refactor: move pipes to their own directories
