+++ src/app/pages/courses-page/pages/courses-list-page/pipes/duration/duration.pipe.ts
