+++ src/app/pages/courses-page/pages/courses-list-page/pipes/filter/filter.pipe.spec.ts
